\name{iForecast.var}
\alias{iForecast.var}

\encoding{latin1}

\title{Produce multistep forecasts from machine learning VAR}

\description{
It generates multistep forecasts of machine learning VAR.
}

\usage{

iForecast.var(object, n.ahead)

}

\arguments{
  \item{object}{
The object generated by \code{tts.var}.
}

  \item{n.ahead}{The number of out-of-sample forecasting periods. If n.ahead=1,
  it is one-step forecast; if n.ahead>1, it computes multistep forecasts by recursive method.
}

}


\details{

This function generates multistep forecasts of machine learning VAR.
}


\author{
Ho Tsung-wu <tsungwu@ntnu.edu.tw>, College of Management, National Taiwan Normal University.

}

\examples{
data(macrodata)
y=timeSeries::as.timeSeries(macrodata[,-1])
VLD=window(y,start="2019-01-01",end=end(y))
#OUT1=tts.var(data=y,
#             p=3,
#             method="enet",
#             train.end="2018-12-01",
#             type=c("none","trend","season","both")[1])

#fcst_ml=iForecast.var(OUT1, n.ahead=nrow(VLD))
}

