\name{data-sets}


\alias{ES_15m}

\alias{ES_Daily}

\alias{macrodata}

\title{Economic and Financial Data Sets}

\encoding{UTF-8}

\description{

ES_15m is 15-min realized absolute variance of e-mini S&P 500. macrodata contains monthly US unemployment(unrate), ES_Daily is daily realized absolute variance of e-mini S&P 500. macrodata contains monthly US unemployment(unrate) and  and year-to-year changes in three regional business cycle indices (OED, NAFTA, and G7).

}


\usage{
data(ES_15m)
data(macrodata)
data(ES_Daily)
}



\value{

    an object of class \code{"zoo"}.

}


\keyword{datasets}

